+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Change
 
 * [#771](https://github.com/xmake-io/xmake/issues/771): Check includedirs, linkdirs and frameworkdirs
+* [#774](https://github.com/xmake-io/xmake/issues/774): Support ltui windows resize for `xmake f --menu`
 
 ## v2.3.3
 
@@ -728,6 +729,7 @@
 ### 改进
 
 * [#771](https://github.com/xmake-io/xmake/issues/771): 检测includedirs,linkdirs和frameworkdirs的输入有效性
+* [#774](https://github.com/xmake-io/xmake/issues/774): `xmake f --menu`可视化配置菜单支持窗口大小Resize调整
 
 ## v2.3.3